projects/Gemini Assistant/app/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
google-genai==1.14.0
google-auth-oauthlib==1.2.1
google-auth==2.40.1
python-dotenv==1.1.0
markdown==3.8
